requirements.txt
asyncpg
fastapi
openai
python-dotenv
uvicorn
